requirements.txt
streamlit 
tensorflow 
matplotlib 
pillow 
numpy 
seaborn 
scikit-learn
google.generativeai